CascadeProjects/windsurf-project/requirements.txt
PyQt6>=6.4.0
paho-mqtt>=2.0.0
pyyaml>=6.0
theming>=1.4.0
